+++ python/ccxt/base/errors.py
@@ -37,7 +37,15 @@ error_hierarchy = {
 
 __all__ = []
 
-from ccxt.base.functions import un_camel_case  # noqa
+import re  # noqa
+
+
+def un_camel_case(string):
+    if re.match(r'^[A-Z0-9_]+$', string):
+        return string
+    first = re.sub(r'([a-z0-9])([A-Z])', r'\1_\2', string)
+    second = re.sub(r'([A-Z0-9])([A-Z0-9][a-z])', r'\1_\2', first)
+    return second.lower()
 
 
 def error_factory(dictionary, super_class):
